Fixing incorrect case in test files
Travis tests failed because of incorrect case on some class names that
couldn’t be autoloaded properly on a case sensitive file system.

Used phpstan to find all cases of incorrect case in the test files and
fix.  No other changes to the tests currently.

Did have to add a couple of files to the dev autoloader for phpstan to
run against the test files.